Add first tests as an example
