+++ src/cookies.test.js
@@ -1,0 +1,9 @@
+import { createCookie, getCookie } from './cookies'
+
+test('getCookie function exists', () => {
+  expect(getCookie).toBeInstanceOf(Function)
+})
+
+test('createCookie function exists', () => {
+  expect(createCookie).toBeInstanceOf(Function)
+})
